separate booking route added
