+++ backend/src/app.js
@@ -14,6 +14,7 @@ import paymentRoutes from "./routes/payment.routes.js";
 import genresRoutes from "./routes/genres.routes.js";
 import seasonTicketsRoutes from "./routes/season-tickets.routes.js";
 import performanceEventsRoutes from "./routes/performanceEvents.routes.js";
+import bookingRoutes from "./routes/booking.routes.js"
 
 const app = express();
 
@@ -34,6 +35,7 @@ app.use("/api/season-tickets", seasonTicketsRoutes);
 app.use("/theaterAdmin", theaterAdmin);
 app.use("/api/genres", genresRoutes);
 app.use("/api/theater", theatersRoutes);
+app.use("/api/booking", bookingRoutes);
 
 app.use(errorHandler);
