+++ backend/src/controllers/booking.controller.js
@@ -1,0 +1,14 @@
+import bookingService from "../services/booking.service.js";
+
+
+const getByUserId = async (req, res, next) => {
+    const { userId } = req.params;
+    try {
+      const response = await bookingService.getByUserId({ userId });
+      res.json(response);
+    } catch (error) {
+      next(error);
+    }
+  };
+
+  export default { getByUserId };
