+++ backend/src/routes/season-tickets.routes.js
@@ -5,7 +5,6 @@ const router = express.Router();
 
 router.get("/", seasonTicketsController.list);
 router.get("/:id", seasonTicketsController.getById);
-router.get("/user/:userId", seasonTicketsController.getByUserId);
 router.post("/", seasonTicketsController.create);
 
 export default router;
