+++ frontend/src/services/booking.service.js
@@ -1,0 +1,12 @@
+import axiosInstance from './axiosInstance';
+
+const getByUserId = async ({ userId }) => {
+    try {
+      const response = await axiosInstance.get(`/api/booking/${userId}`);
+      return response.data;
+    } catch (error) {
+      throw error.response ? error.response.data : error;
+    }
+  };
+
+  export default { getByUserId };
